+++ predicate/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod delete_expr;
+pub mod delete_predicate;
 pub mod predicate;
 pub mod regex;
 pub mod serialize;
