+++ predicate/src/serialize.rs
@@ -12,19 +12,11 @@ use data_types::timestamp::TimestampRange;
 use generated_types::influxdata::iox::catalog::v1 as proto;
 use snafu::{ResultExt, Snafu};
 
-use crate::{delete_expr::DeleteExpr, predicate::Predicate};
+use crate::{delete_expr::DeleteExpr, delete_predicate::DeletePredicate};
 
-#[derive(Debug, Snafu)]
-pub enum SerializeError {
-    #[snafu(display("cannot convert datafusion expr: {}", source))]
-    CannotConvertDataFusionExpr {
-        source: crate::delete_expr::DataFusionToExprError,
-    },
-}
-
-/// Serialize IOx [`Predicate`] to a protobuf object.
-pub fn serialize(predicate: &Predicate) -> Result<proto::Predicate, SerializeError> {
-    let proto_predicate = proto::Predicate {
+/// Serialize IOx [`DeletePredicate`] to a protobuf object.
+pub fn serialize(predicate: &DeletePredicate) -> proto::Predicate {
+    proto::Predicate {
         table_names: serialize_optional_string_set(&predicate.table_names),
         field_columns: serialize_optional_string_set(&predicate.field_columns),
         partition_key: serialize_optional_string(&predicate.partition_key),
@@ -32,16 +24,9 @@ pub fn serialize(predicate: &Predicate) -> Result<proto::Predicate, SerializeErr
         exprs: predicate
             .exprs
             .iter()
-            .map(|expr| {
-                let expr: DeleteExpr = expr
-                    .clone()
-                    .try_into()
-                    .context(CannotConvertDataFusionExpr)?;
-                Ok(expr.into())
-            })
-            .collect::<Result<Vec<proto::Expr>, SerializeError>>()?,
-    };
-    Ok(proto_predicate)
+            .map(|expr| expr.clone().into())
+            .collect(),
+    }
 }
 
 fn serialize_optional_string_set(
@@ -73,8 +58,10 @@ pub enum DeserializeError {
 }
 
 /// Deserialize IOx [`Predicate`] from a protobuf object.
-pub fn deserialize(proto_predicate: &proto::Predicate) -> Result<Predicate, DeserializeError> {
-    let predicate = Predicate {
+pub fn deserialize(
+    proto_predicate: &proto::Predicate,
+) -> Result<DeletePredicate, DeserializeError> {
+    let predicate = DeletePredicate {
         table_names: deserialize_optional_string_set(&proto_predicate.table_names),
         field_columns: deserialize_optional_string_set(&proto_predicate.field_columns),
         partition_key: deserialize_optional_string(&proto_predicate.partition_key),
@@ -84,9 +71,9 @@ pub fn deserialize(proto_predicate: &proto::Predicate) -> Result<Predicate, Dese
             .iter()
             .map(|expr| {
                 let expr: DeleteExpr = expr.clone().try_into().context(CannotDeserializeExpr)?;
-                Ok(expr.into())
+                Ok(expr)
             })
-            .collect::<Result<Vec<datafusion::logical_plan::Expr>, DeserializeError>>()?,
+            .collect::<Result<Vec<DeleteExpr>, DeserializeError>>()?,
     };
     Ok(predicate)
 }
@@ -110,7 +97,7 @@ fn deserialize_timestamp_range(r: &Option<proto::TimestampRange>) -> Option<Time
 
 #[cfg(test)]
 mod tests {
-    use crate::predicate::{ParseDeletePredicate, PredicateBuilder};
+    use crate::delete_predicate::ParseDeletePredicate;
 
     use super::*;
 
@@ -118,12 +105,12 @@ mod tests {
     fn test_roundtrip() {
         let table_name = "my_table";
         let predicate = delete_predicate(table_name);
-        let proto = serialize(&predicate).unwrap();
+        let proto = serialize(&predicate);
         let recovered = deserialize(&proto).unwrap();
         assert_eq!(predicate, recovered);
     }
 
-    fn delete_predicate(table_name: &str) -> Predicate {
+    fn delete_predicate(table_name: &str) -> DeletePredicate {
         let start_time = "11";
         let stop_time = "22";
         let predicate = r#"city=Boston and cost!=100 and temp=87.5 and good=true"#;
@@ -131,14 +118,8 @@ mod tests {
         let parse_delete_pred =
             ParseDeletePredicate::try_new(start_time, stop_time, predicate).unwrap();
 
-        let mut del_predicate_builder = PredicateBuilder::new()
-            .table(table_name)
-            .timestamp_range(parse_delete_pred.start_time, parse_delete_pred.stop_time);
-
-        for expr in parse_delete_pred.predicate {
-            del_predicate_builder = del_predicate_builder.add_expr(expr);
-        }
-
-        del_predicate_builder.build()
+        let mut pred: DeletePredicate = parse_delete_pred.into();
+        pred.table_names = Some(IntoIterator::into_iter([table_name.to_string()]).collect());
+        pred
     }
 }
