+++ server/src/db/chunk.rs
@@ -18,7 +18,10 @@ use mutable_buffer::chunk::snapshot::ChunkSnapshot;
 use observability_deps::tracing::debug;
 use parquet_file::chunk::ParquetChunk;
 use partition_metadata::TableSummary;
-use predicate::predicate::{Predicate, PredicateMatch};
+use predicate::{
+    delete_predicate::DeletePredicate,
+    predicate::{Predicate, PredicateMatch},
+};
 use query::{exec::stringset::StringSet, QueryChunk, QueryChunkMeta};
 use read_buffer::RBChunk;
 use snafu::{OptionExt, ResultExt, Snafu};
@@ -546,7 +549,7 @@ impl QueryChunkMeta for DbChunk {
     }
 
     // return a reference to delete predicates of the chunk
-    fn delete_predicates(&self) -> &[Arc<Predicate>] {
+    fn delete_predicates(&self) -> &[Arc<DeletePredicate>] {
         let pred = &self.meta.delete_predicates;
         debug!(?pred, "Delete predicate in  DbChunk");
